% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basalArea.R
\name{basalArea}
\alias{basalArea}
\title{Calculate basal area from DBH}
\usage{
basalArea(x)
}
\arguments{
\item{x}{DBH (diameter at breast height) input value to calculate basal area}
}
\value{
Returns the basal area from input DBH
}
\description{
Calculate basal area from DBH
}
\examples{
DBH <- c(30, 50, 70, 90)
basalArea(DBH)


}
